--- a/jwe/internal/cmd/genheader/main.go
+++ b/jwe/internal/cmd/genheader/main.go
@@ -475,12 +475,9 @@ func generateHeaders() error {
 	fmt.Fprintf(&buf, "\n}")
 
 	fmt.Fprintf(&buf, "\n\nfunc (h stdHeaders) MarshalJSON() ([]byte, error) {")
-	fmt.Fprintf(&buf, "\nctx, cancel := context.WithCancel(context.Background())")
-	fmt.Fprintf(&buf, "\ndefer cancel()")
 	fmt.Fprintf(&buf, "\ndata := make(map[string]interface{})")
 	fmt.Fprintf(&buf, "\nfields := make([]string, 0, %d)", len(fields))
-	fmt.Fprintf(&buf, "\nfor iter := h.Iterate(ctx); iter.Next(ctx); {")
-	fmt.Fprintf(&buf, "\npair := iter.Pair()")
+	fmt.Fprintf(&buf, "\nfor _, pair := range h.makePairs() {")
 	fmt.Fprintf(&buf, "\nfields = append(fields, pair.Key.(string))")
 	fmt.Fprintf(&buf, "\ndata[pair.Key.(string)] = pair.Value")
 	fmt.Fprintf(&buf, "\n}")