--- a/jwk/ecdsa_gen.go
+++ b/jwk/ecdsa_gen.go
@@ -555,12 +555,9 @@ LOOP:
 }
 
 func (h ecdsaPrivateKey) MarshalJSON() ([]byte, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	data := make(map[string]interface{})
 	fields := make([]string, 0, 12)
-	for iter := h.Iterate(ctx); iter.Next(ctx); {
-		pair := iter.Pair()
+	for _, pair := range h.makePairs() {
 		fields = append(fields, pair.Key.(string))
 		data[pair.Key.(string)] = pair.Value
 	}
@@ -1118,12 +1115,9 @@ LOOP:
 }
 
 func (h ecdsaPublicKey) MarshalJSON() ([]byte, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	data := make(map[string]interface{})
 	fields := make([]string, 0, 11)
-	for iter := h.Iterate(ctx); iter.Next(ctx); {
-		pair := iter.Pair()
+	for _, pair := range h.makePairs() {
 		fields = append(fields, pair.Key.(string))
 		data[pair.Key.(string)] = pair.Value
 	}