--- a/jwk/okp_gen.go
+++ b/jwk/okp_gen.go
@@ -523,12 +523,9 @@ LOOP:
 }
 
 func (h okpPrivateKey) MarshalJSON() ([]byte, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	data := make(map[string]interface{})
 	fields := make([]string, 0, 11)
-	for iter := h.Iterate(ctx); iter.Next(ctx); {
-		pair := iter.Pair()
+	for _, pair := range h.makePairs() {
 		fields = append(fields, pair.Key.(string))
 		data[pair.Key.(string)] = pair.Value
 	}
@@ -1056,12 +1053,9 @@ LOOP:
 }
 
 func (h okpPublicKey) MarshalJSON() ([]byte, error) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 	data := make(map[string]interface{})
 	fields := make([]string, 0, 10)
-	for iter := h.Iterate(ctx); iter.Next(ctx); {
-		pair := iter.Pair()
+	for _, pair := range h.makePairs() {
 		fields = append(fields, pair.Key.(string))
 		data[pair.Key.(string)] = pair.Value
 	}