--- a/jwk/symmetric.go
+++ b/jwk/symmetric.go
@@ -1,7 +1,6 @@
 package jwk
 
 import (
-	"context"
 	"crypto"
 	"fmt"
 
@@ -51,8 +50,7 @@ func (k *symmetricKey) Thumbprint(hash crypto.Hash) ([]byte, error) {
 func (k *symmetricKey) PublicKey() (Key, error) {
 	newKey := NewSymmetricKey()
 
-	for iter := k.Iterate(context.TODO()); iter.Next(context.TODO()); {
-		pair := iter.Pair()
+	for _, pair := range k.makePairs() {
 		if err := newKey.Set(pair.Key.(string), pair.Value); err != nil {
 			return nil, errors.Wrapf(err, `failed to set field %s`, pair.Key)
 		}