Iteration of lestrrat-go/iter using goroutine and channel is sometimes useful with a generic interface, but it can be slow.

I tried the benchmark for `LookupKey`, it has been improved.

```go
package jwk_test

import (
	"context"
	"testing"

	"github.com/lestrrat-go/jwx/jwk"
)

const u = "https://www.googleapis.com/oauth2/v3/certs"

func BenchmarkVerifySet(b *testing.B) {
	ctx := context.Background()
	ar := jwk.NewAutoRefresh(ctx)
	ar.Configure(u)
	set, err := ar.Fetch(ctx, u)
	if err != nil {
		b.Fatal(err)
	}
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, ok := set.LookupKeyID("REPLACE_THE_REAL_KEY_ID")
		if !ok {
			b.Fatal("no key")
		}
	}
}
```


before
```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/jwk
cpu: 11th Gen Intel(R) Core(TM) i3-1115G4 @ 3.00GHz
BenchmarkVerifySet-4   	 1214281	       982.7 ns/op	     304 B/op	       7 allocs/op
PASS
ok  	github.com/lestrrat-go/jwx/jwk	2.028s
```

after
```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/jwk
cpu: 11th Gen Intel(R) Core(TM) i3-1115G4 @ 3.00GHz
BenchmarkVerifySet-4   	14858936	        71.80 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/lestrrat-go/jwx/jwk	1.296s
```