@johejo 2021-09-17
Oops, my benchmark func name is not good...  should be BenchmarkLookupKey....